Fix #622
- When object width/height are changed from the inspector
we shouldn't trigger resize events from the node-editing
extension since this is not an action that belongs to it.
Node-editing extension should be concerned with actions
where grapples/cues are used for resizing, and node movement.

- The Chise resizeNodes function already handles undo, we don't need
to trigger resize extension events to get undo functionality. In fact,
this was causing the "undo that does nothing" mentioned as a reply to #622.